# backend-app/configs/swagger_config.json
{
    "swagger": "2.0",
    "info": {
        "title": "Software Voting Machine API",
        "description": "This API allows us to manage all related endpoints to our software voting machine app.",
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/options": {
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Create a new option",
                "description": "Creates a new option with the given data.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "option",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OptionCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Option created successfully",
                        "schema": {
                            "$ref": "#/definitions/OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get all options",
                "description": "Returns a list of all options in the database.",
                "responses": {
                    "200": {
                        "description": "A list of options",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/OptionResponse"
                            }
                        }
                    }
                }
            }
        },
        "/options/{option_id}": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get an option by ID",
                "description": "Retrieves the option with the specified ID.",
                "parameters": [
                    {
                        "name": "option_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option found",
                        "schema": {
                            "$ref": "#/definitions/OptionResponse"
                        }
                    },
                    "404": {
                        "description": "Option not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Options"
                ],
                "summary": "Update an option by ID",
                "description": "Updates the option with the given ID using the provided data.",
                "parameters": [
                    {
                        "name": "option_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "option",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OptionResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option updated successfully",
                        "schema": {
                            "$ref": "#/definitions/OptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Option not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Options"
                ],
                "summary": "Delete an option by ID",
                "description": "Deletes the option with the specified ID.",
                "parameters": [
                    {
                        "name": "option_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option deleted successfully"
                    },
                    "404": {
                        "description": "Option not found"
                    }
                }
            }
        },
...
        "/profile_statuses": {
            "post": {
                "tags": [
                    "ProfileStatuses"
                ],
                "summary": "Create a new profile status",
                "description": "Creates a new profile status with the given data.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "profile_status",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProfileStatusCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Profile status created successfully",
                        "schema": {
                            "$ref": "#/definitions/ProfileStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "ProfileStatuses"
                ],
                "summary": "Get all profile statuses",
                "description": "Returns a list of all profile statuses in the database.",
                "responses": {
                    "200": {
                        "description": "A list of profile statuses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProfileStatusResponse"
                            }
                        }
                    }
                }
            }
        },
        "/profile_statuses/{id}": {
            "get": {
                "tags": [
                    "ProfileStatuses"
                ],
                "summary": "Get a profile status by ID",
                "description": "Retrieves the profile status with the specified ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile status found",
                        "schema": {
                            "$ref": "#/definitions/ProfileStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Profile status not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProfileStatuses"
                ],
                "summary": "Update a profile status by ID",
                "description": "Updates the profile status with the given ID using the provided data.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "profile_status",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProfileStatusCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/ProfileStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Profile status not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProfileStatuses"
                ],
                "summary": "Delete a profile status by ID",
                "description": "Deletes the profile status with the specified ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile status deleted successfully"
                    },
                    "404": {
                        "description": "Profile status not found"
                    }
                }
            }
        },
        "/questions": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question",
                "description": "Creates a new question with the given data.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "question",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/QuestionCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created successfully",
                        "schema": {
                            "$ref": "#/definitions/QuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get all questions",
                "description": "Returns a list of all questions in the database.",
                "responses": {
                    "200": {
                        "description": "A list of questions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/QuestionResponse"
                            }
                        }
                    }
                }
            }
        },
        "/questions/{question_id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get a question by ID",
                "description": "Retrieves the question with the specified ID.",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question found",
                        "schema": {
                            "$ref": "#/definitions/QuestionResponse"
                        }
                    },
                    "404": {
                        "description": "Question not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update a question by ID",
                "description": "Updates the question with the given ID using the provided data.",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "question",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/QuestionCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated successfully",
                        "schema": {
                            "$ref": "#/definitions/QuestionResponse"
                        }
                    },
                    "404": {
                        "description": "Question not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a question by ID",
                "description": "Deletes the question with the specified ID.",
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully"
                    },
                    "404": {
                        "description": "Question not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "OptionCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the option"
                },
                "question_id": {
                    "type": "integer",
                    "description": "The ID of the question associated with the option"
                },
                "session_id": {
                    "type": "integer",
                    "description": "The ID of the session associated with the option"
                }
            },
            "required": [
                "name",
                "question_id",
                "session_id"
            ]
        },
        "OptionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "The unique identifier of the option"
                },
                "name": {
                    "type": "string",
                    "description": "The name of the option"
                },
                "question_id": {
                    "type": "integer",
                    "description": "The ID of the question associated with the option"
                },
                "session_id": {
                    "type": "integer",
                    "description": "The ID of the session associated with the option"
                }
            }
        },
...
        "ProfileStatusCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The status of the profile"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date",
                    "description": "The date the profile status was last updated"
                }
            },
            "required": [
                "user_id",
                "name"
            ]
        },
        "ProfileStatusResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "The ID of the profile status"
                },
                "name": {
                    "type": "string",
                    "description": "The status of the profile"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date",
                    "description": "The date the profile status was last updated"
                }
            }
        },
        "QuestionCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the question"
                },
                "voting_session_id": {
                    "type": "integer",
                    "description": "The ID of the voting session associated with the question"
                }
            },
            "required": [
                "name",
                "voting_session_id"
            ]
        },
        "QuestionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "The unique identifier of the question"
                },
                "name": {
                    "type": "string",
                    "description": "The name of the question"
                },
                "voting_session_id": {
                    "type": "integer",
                    "description": "The ID of the voting session associated with the question"
                }
            }
        },
...
    }
}